+++ src/ir_LG.cpp
@@ -19,7 +19,7 @@
 //+=============================================================================
 #if DECODE_LG
 bool IRrecv::decodeLG() {
-    int offset = 1; // Skip first space
+    unsigned int offset = 1; // Skip first space
 
     // Check we have the right amount of data  +3 for start bit mark and space + stop bit mark
     if (results.rawlen <= (2 * LG_BITS) + 3)